+++ src/graphic/Fast3D/gfx_direct3d11.h
@@ -10,7 +10,7 @@
 #include <imgui.h>
 
 extern struct GfxRenderingAPI gfx_direct3d11_api;
-ImTextureID gfx_d3d11_get_texture_by_id(int id);
+ImTextureID gfx_d3d11_get_texture_by_id(unsigned long long id);
 
 #endif
