+++ src/graphic/Fast3D/gfx_metal.cpp
@@ -1295,7 +1295,7 @@ FilteringMode gfx_metal_get_texture_filter() {
     return mctx.current_filter_mode;
 }
 
-ImTextureID gfx_metal_get_texture_by_id(int fb_id) {
+ImTextureID gfx_metal_get_texture_by_id(unsigned long long fb_id) {
     return (void*)mctx.textures[fb_id].texture;
 }
